utf-8 decoding of input streams

+++ source/antlr/v4/runtime/ANTLRErrorListener.d
@@ -96,7 +96,7 @@ interface ANTLRErrorListener(U, V)
      * @param configs the ATN configuration set where the ambiguity was
      * identified
      */
-    public void reportAmbiguity(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportAmbiguity(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
                                 bool exact, BitSet ambigAlts, ATNConfigSet configs);
 
     /**
@@ -122,7 +122,7 @@ interface ANTLRErrorListener(U, V)
      * @param configs the ATN configuration set where the SLL conflict was
      * detected
      */
-    public void reportAttemptingFullContext(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportAttemptingFullContext(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
                                             BitSet conflictingAlts, ATNConfigSet configs);
 
     /**
@@ -163,7 +163,7 @@ interface ANTLRErrorListener(U, V)
      * @param configs the ATN configuration set where the unambiguous prediction
      * was determined
      */
-    public void reportContextSensitivity(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportContextSensitivity(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
                                          int prediction, ATNConfigSet configs);
 
 }
