+++ source/antlr/v4/runtime/BaseErrorListener.d
@@ -21,17 +21,17 @@ class BaseErrorListener(U, V) : ANTLRErrorListener!(U, V)
     {
     }
 
-    public void reportAmbiguity(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportAmbiguity(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
         bool exact, BitSet ambigAlts, ATNConfigSet configs)
     {
     }
 
-    public void reportAttemptingFullContext(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportAttemptingFullContext(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
         BitSet conflictingAlts, ATNConfigSet configs)
     {
     }
 
-    public void reportContextSensitivity(InterfaceParser recognizer, DFA dfa, int startIndex, int stopIndex,
+    public void reportContextSensitivity(InterfaceParser recognizer, DFA dfa, size_t startIndex, size_t stopIndex,
         int prediction, ATNConfigSet configs)
     {
     }
